There's a lot, buckle up. q's make a QuestionList
So. I tried making the questions update firestore which
worked but then it was non functional from there because
firestore has diddly for functonality. so there goes many many hours of
work. but I found a solution. it's jank, but now questions
delete things from a list in local storage, and your listed options are
whatever is left over